+++ integration/mip/module_tasks/legend_item_segment_task_3.py
@@ -1,9 +1,9 @@
 # Copyright 2024 InferLink Corporation
 
-from tasker.utils.docker_task import DockerTask
-from tasker.module_tasks.text_spotting_task_6 import TextSpottingTask6
-from tasker.utils.checker import check_file_exists
-from tasker.module_tasks.registry import register_task
+from mip.utils.docker_task import DockerTask
+from mip.module_tasks.text_spotting_task_6 import TextSpottingTask6
+from mip.utils.checker import check_file_exists
+from mip.module_tasks.registry import register_task
 
 
 @register_task
