+++ integration/mip/module_tasks/legend_segment_task_2.py
@@ -1,9 +1,9 @@
 # Copyright 2024 InferLink Corporation
 
-from tasker.utils.docker_task import DockerTask
-from tasker.module_tasks.start_task_0 import StartTask0
-from tasker.utils.checker import check_file_exists
-from tasker.module_tasks.registry import register_task
+from mip.utils.docker_task import DockerTask
+from mip.module_tasks.start_task_0 import StartTask0
+from mip.utils.checker import check_file_exists
+from mip.module_tasks.registry import register_task
 
 
 @register_task
