+++ integration/mip/module_tasks/line_extract_task_7.py
@@ -1,8 +1,8 @@
 # Copyright 2024 InferLink Corporation
 
-from tasker.utils.docker_task import DockerTask
-from tasker.module_tasks.legend_item_description_task_4 import LegendItemDescriptionTask4
-from tasker.module_tasks.registry import register_task
+from mip.utils.docker_task import DockerTask
+from mip.module_tasks.legend_item_description_task_4 import LegendItemDescriptionTask4
+from mip.module_tasks.registry import register_task
 
 
 @register_task
