+++ integration/mipper.py
@@ -5,10 +5,10 @@ import sys
 
 import luigi
 
-from tasker.module_tasks.registry import registry_lookup, get_task_names
-from tasker.module_tasks import *
-from tasker.utils.config import Config
-from tasker.utils.options import Options
+from mip.module_tasks.registry import registry_lookup, get_task_names
+from mip.module_tasks import *
+from mip.utils.config import Config
+from mip.utils.options import Options
 
 # datetime.datetime.now().strftime("%H.%M.%S")
